semana12/microdesafio/consumer/requirements.txt
kafka-python
quixstreams
